requirements.txt
requests==2.32.3
python-dotenv==1.1.1
python-telegram-bot==21.6
bs4==0.0.2
ua-generator==1.0.6